# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RunPG API",
    "description": "API for the RunPG mobile game and watch app",
    "version": "1.0.0"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/user/": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users successfully obtained"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "uid": {
                  "example": "any"
                },
                "heroClass": {
                  "example": "any"
                }
              }
            },
            "description": "new user",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User created"
          },
          "400": {
            "description": "Request is not valid"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/name/{name}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by name",
        "description": "",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "404": {
            "description": "Could not find the user"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by id",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User found"
          },
          "400": {
            "description": "Wrong user id"
          },
          "404": {
            "description": "Could not find the user"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted"
          },
          "400": {
            "description": "Request is not valid"
          },
          "404": {
            "desctiption": "User not found",
            "description": "Not Found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/friend/{friendId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a friend",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "friendId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend found"
          },
          "400": {
            "description": "Friend not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/friend": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all the friends",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Friends found"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/friend/{id}": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Add a new friend",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Friend added"
          },
          "400": {
            "description": "Bad userId or id"
          },
          "409": {
            "description": "This friend is already added"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/xp": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Modify user xp",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "xp",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "xp": {
                  "example": "any"
                }
              }
            },
            "description": "new user xp",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User xp updated"
          },
          "400": {
            "description": "Bad userId or xp"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/notification": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all user notifications",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications found"
          },
          "400": {
            "description": "Bad userId"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/notification/{type}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all user notifications by type",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications found"
          },
          "400": {
            "description": "Bad userId or type"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/notification/{type}/{senderId}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all user notifications by type",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "senderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications found"
          },
          "400": {
            "description": "Bad userId or type"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a notification",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "senderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification created"
          },
          "400": {
            "description": "Bad userId, type or senderId"
          },
          "409": {
            "description": "Notification already exists"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a notification",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "senderId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Notification deleted"
          },
          "400": {
            "description": "Bad userId, type or senderId"
          },
          "404": {
            "description": "Notification does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/join/{guildId}": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Join the guild",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "guildId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully added to guild"
          },
          "400": {
            "description": "Bad userId or guildId"
          },
          "404": {
            "description": "User or guild does not exists"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/inventory/equipement": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Give an equipement to an user",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "equipementBaseId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Equipement successfully added to user inventory"
          },
          "400": {
            "description": "Bad userId or equipementBaseId"
          },
          "404": {
            "description": "User or equipement base does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/inventory/item": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Give item(s) to an user",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "itemId": {
                  "example": "any"
                },
                "stackSize": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Items(s) successfully added to user inventory"
          },
          "400": {
            "description": "Bad userId, itemId or stackSize"
          },
          "404": {
            "description": "User or item does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/character": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a character of an user",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User character found"
          },
          "400": {
            "description": "Bad userId"
          },
          "404": {
            "description": "User does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
    "/user/{userId}/inventory": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get the inventory of an user",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User inventoy found"
          },
          "400": {
            "description": "userId is not valid"
          },
          "404": {
            "description": "User does not exist"
          },
          "500": {
            "description": "Server encountered an internal error"
          }
        }
      }
    },
...
  }
}